+++ contracts/orderbook/OrderBook.sol
@@ -536,7 +536,7 @@ contract OrderBook is
             // The state changes produced here are handled in _recordVaultIO so
             // that local storage writes happen before writes on the interpreter.
             StateNamespace namespace_ = StateNamespace.wrap(
-                uint(uint160(order_.owner))
+                uint256(uint160(order_.owner))
             );
             (uint256[] memory stack_, uint256[] memory kvs_) = order_
                 .evaluable
